backend/requirements-dev.txt
amqp==5.3.1
asgiref==3.10.0
astroid==4.0.1
bandit==1.8.6
billiard==4.2.2
brotli==1.2.0
celery==5.5.3
cffi==2.0.0
docx2pdf
cfgv==3.4.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
config==0.5.1
cssselect2==0.8.0
cryptography>=41.0.0
dill==0.4.0
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
django-crontab==0.7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docxtpl==0.20.2
drf-yasg==1.21.11
et_xmlfile==2.0.0
filelock==3.20.0
flake8==7.3.0
fonttools==4.60.1
identify==2.6.15
inflection==0.5.1
isort==7.0.0
Jinja2==3.1.6
kombu==5.5.4
load-dotenv==0.1.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mysqlclient==2.2.7
nodeenv==1.9.1
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
pycodestyle==2.14.0
pycparser==2.23
pydyf==0.11.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.2
PyMySQL==1.1.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
pikepdf==8.4.0
rich==14.2.0
six==1.17.0
sqlparse==0.5.3
stevedore==5.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
virtualenv==20.35.3
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
pypandoc
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-redis==6.0.0
zipfile36